# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCQ Online Test Platform API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/examiner/users/examiners": {
      "post": {
        "tags": [
          "examiner-users"
        ],
        "summary": "Create an examiner account",
        "description": "Create a new examiner account (admin only).\n\nThe password must satisfy the platform policy (minimum 12 characters).\nBy default the new examiner must change their password on first login.\nRequires an admin examiner session and the `X-CSRF-Token` header.",
        "operationId": "create_examiner_api_v1_examiner_users_examiners_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExaminerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created examiner account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ExaminerUserOut_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "409": {
            "description": "An examiner with this username already exists.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The resource was modified by another request.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/examiner/users/examinees/{user_id}/erase": {
      "post": {
        "tags": [
          "examiner-users"
        ],
        "summary": "Request GDPR erasure of an examinee",
        "description": "Request asynchronous erasure of an examinee's personal data (GDPR Art. 17).\n\nReturns **202** with a `job_id`; the erasure is processed in the\nbackground. Requires an admin examiner session and the `X-CSRF-Token`\nheader.",
        "operationId": "erase_examinee_api_v1_examiner_users_examinees__user_id__erase_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Erasure job accepted for asynchronous processing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_EraseExamineeResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/examiner/settings": {
      "get": {
        "tags": [
          "examiner-settings"
        ],
        "summary": "Get platform settings",
        "description": "Read the platform settings (retention, retake policy, question selection mode).\n\nAvailable to any authenticated examiner.",
        "operationId": "get_settings_api_v1_examiner_settings_get",
        "responses": {
          "200": {
            "description": "The current platform settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PlatformSettingsOut_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "patch": {
        "tags": [
          "examiner-settings"
        ],
        "summary": "Update platform settings",
        "description": "Partially update platform settings; omitted fields are left unchanged.\n\nRequires an admin examiner session and the `X-CSRF-Token` header.",
        "operationId": "patch_settings_api_v1_examiner_settings_patch",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformSettingsPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The settings after applying the update.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PlatformSettingsOut_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateExaminerRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "description": "Username for the new examiner (whitespace trimmed).",
            "examples": [
              "examiner2"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "description": "Initial password; must satisfy the platform password policy (minimum 12 characters)."
          },
          "force_password_change": {
            "type": "boolean",
            "title": "Force Password Change",
            "description": "When true, the examiner must change the password on first login.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "CreateExaminerRequest",
        "description": "Payload to create a new examiner account (admin only)."
      },
...
      "EraseExamineeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "The examinee whose data will be erased."
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id",
            "description": "Identifier of the asynchronous erasure job."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Job status at acceptance time.",
            "default": "accepted",
            "examples": [
              "accepted"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "job_id"
        ],
        "title": "EraseExamineeResponse",
        "description": "Acknowledgement of an accepted GDPR erasure request."
      },
...
      "ExaminerUserOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Unique identifier of the examiner."
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Examiner username.",
            "examples": [
              "examiner2"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the account is active."
          },
          "force_password_change": {
            "type": "boolean",
            "title": "Force Password Change",
            "description": "Whether the examiner must change their password on first login."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the account was created (UTC)."
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "is_active",
          "force_password_change",
          "created_at"
        ],
        "title": "ExaminerUserOut",
        "description": "A created examiner account."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PlatformSettingsOut": {
        "properties": {
          "retention_days_completed": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Retention Days Completed",
            "description": "Days completed sessions are retained before automatic deletion.",
            "examples": [
              365
            ]
          },
          "allow_examinee_retake": {
            "type": "boolean",
            "title": "Allow Examinee Retake",
            "description": "Whether examinees who completed the test may take it again."
          },
          "question_selection_mode": {
            "$ref": "#/components/schemas/SelectionMode",
            "description": "How questions are selected when a new test session starts."
          }
        },
        "type": "object",
        "required": [
          "retention_days_completed",
          "allow_examinee_retake",
          "question_selection_mode"
        ],
        "title": "PlatformSettingsOut",
        "description": "Current platform settings."
      },
      "PlatformSettingsPatchRequest": {
        "properties": {
          "retention_days_completed": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3650.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days Completed",
            "description": "New retention period in days (1-3650)."
          },
          "allow_examinee_retake": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Examinee Retake",
            "description": "New retake policy."
          },
          "question_selection_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionMode"
              },
              {
                "type": "null"
              }
            ],
            "description": "New question selection mode."
          }
        },
        "type": "object",
        "title": "PlatformSettingsPatchRequest",
        "description": "Partial settings update; omitted fields are left unchanged."
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Unique request identifier; matches the X-Request-Id response header.",
            "examples": [
              "9f1b2c3d4e5f6a7b"
            ]
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "description": "Per-request metadata included in every success envelope."
      },
...
      "SelectionMode": {
        "type": "string",
        "enum": [
          "FIXED_ORDER",
          "RANDOM_SAMPLE"
        ],
        "title": "SelectionMode"
      },
...
      "SuccessResponse_EraseExamineeResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EraseExamineeResponse",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[EraseExamineeResponse]"
      },
...
      "SuccessResponse_ExaminerUserOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExaminerUserOut",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExaminerUserOut]"
      },
      "SuccessResponse_PlatformSettingsOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlatformSettingsOut",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[PlatformSettingsOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "FieldError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "description": "Dotted path of the invalid request field.",
            "examples": [
              "username"
            ],
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable description of the validation failure.",
            "examples": [
              "String should have at least 1 character"
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "ProblemDetails": {
        "description": "RFC 7807 problem details returned with content type application/problem+json.",
        "properties": {
          "type": {
            "description": "URI identifying the problem type.",
            "examples": [
              "https://api.example.com/problems/unauthenticated"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "examples": [
              "Unauthenticated"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "examples": [
              401
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation specific to this occurrence.",
            "examples": [
              "Authentication required."
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path on which the problem occurred.",
            "examples": [
              "/api/v1/auth/me"
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request identifier for log correlation; matches X-Request-Id.",
            "examples": [
              "9f1b2c3d4e5f6a7b"
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Field-level validation errors (only on 400 validation failures).",
            "title": "Errors"
          },
          "locked_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "When a locked account becomes available again (only on 423 responses).",
            "title": "Locked Until"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Opaque server-side session token issued by the login/entry endpoints as an HttpOnly cookie. Examiner sessions last 1 hour, examinee sessions 4 hours by default."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound CSRF token required on mutating requests. Obtain it from GET /auth/me (or GET /auth/csrf before login)."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}